+++ tests/models/base/models.py
@@ -1,0 +1,16 @@
+#:coding=utf-8:
+
+from beproud.django.commons.models import (
+    DatedModel as DatedModelBase,
+    BaseModel as BaseModelBase
+)
+
+
+class DatedModel(DatedModelBase):
+    class Meta:
+        app_label = 'base'
+
+
+class BaseModel(BaseModelBase):
+    class Meta:
+        app_label = 'base'