+++ tests/testapp/urls.py
@@ -2,7 +2,7 @@
 
 from django.conf.urls import url, include
 
-from beproud.django.commons.tests.test_views import TestViews
+from .views import TestViews
 
 urlpatterns = [
     url(r'', include(TestViews().urls)),